qtcollider: get rid of many floats in favour of doubles

+++ QtCollider/widgets/QcAbstractStepValue.cpp
@@ -23,7 +23,7 @@
 
 #include "QcAbstractStepValue.h"
 
-bool QcAbstractStepValue::modifyStep( float *pStep )
+bool QcAbstractStepValue::modifyStep( double *pStep )
 {
   Qt::KeyboardModifiers mods = QApplication::keyboardModifiers();
   if( !mods ) return false;
